pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "watchlock"
version = "0.1.0"
description = "BLE proximity-based PC lock/unlock daemon"
requires-python = ">=3.11"
dependencies = [
    "bleak>=0.21",
    "pyyaml>=6.0",
    "click>=8.0",

]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff",
]

[project.scripts]
watchlock = "unlock_pc.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
